Created listing 02.30 and associated test
